Support for scalar default values (#27)
* Skip writing of default scalar values

When writing a scalar field, check if the value is equal to the default
value for its type (e.g. "" for string, 0 for numbers). If so, skip the
field.

We have to be careful *not* to skip list elements and map elements even
if they have the default value.

Also when we implement support for oneof fields, we need to be careful
of those too. From the docs: "If you set a oneof field to the default
value (such as setting an int32 oneof field to 0), the "case" of that
oneof field will be set, and the value will be serialized on the wire."

* Handle missing fields by replacing with default values when reading

The round-trip spec is currently failing, but I need to decide what the
correct behaviour is and adjust either the test or the implementation.

* Update RoundTripSpec

Updated the test to account for the semantics of default values and
optional fields.

Instead of expecting an identical message after a round-trip, we define
an equivalence relation that we expect to hold.

* Add unit tests for reading missing fields

* Fix test compile error on Scala 2.11

* Update key-value pair writer behaviour

For consistency with the corresponding reader