Support oneof fields (#30)
* Support for writing coproducts as oneof fields

* Always write oneof fields

Don't skip writing a oneof field even if it has a default scalar value.

The implementation for this is a bit nasty. I added a boolean parameter
to the signature of `PBFieldWriter` that is only used when writing a
scalar field and is ignored in all other cases. Will have a think about
a cleaner solution.

* Support for reading oneof fields

* Require Coproducts (oneof fields) to be wrapped in Option

This is so we can set the field to None when it is missing when reading
a message from protobuf. If it was a plain coproduct type (e.g.
`Int :+: String :+: CNil`), there is no sensible default value we can
set it to.  We could arbitrarily set it to `Inr(0)`, but we would lose
the information that the field was missing from the message.

* Document new features in readme

* Turn off annoying unused import warnings in REPL

* Update class name in readme

* Remove commented-out code

* Fix typo in readme